### 2019-05-15 #1449 @whedon
[PRE REVIEW]: WRF-CMake: integrating CMake support into the Advanced Research WRF (ARW) modelling system
**Submitting author:** @letmaik (<a href="http://orcid.org/0000-0003-3299-1382">Maik Riechert</a>)
**Repository:** <a href="https://github.com/WRF-CMake/WRF" target ="_blank">https://github.com/WRF-CMake/WRF</a>
**Version:** 4
