### 2019-05-15 #1449 @arfon
:wave: @letmaik - many thanks for your submission to JOSS. Please review our [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and prepare a `paper.md` file (and references in a bibtex file) and add them to the repository.
