### 2019-05-15 #1449 @letmaik
@arfon Is it possible to submit papers to JOSS without affiliation?
