### 2019-05-15 #1449 @letmaik
@arfon We changed layout slightly and also replaced affiliation by "Independent scholar" which makes more sense than "None", see also https://journals.ametsoc.org/doi/full/10.1175/BAMS-D-12-00171.1 where they've done the same.
