### 2019-05-15 #1449 @dmey
@letmaik There is still a minor issue with the last paragraph on page 2 (it needs to be moved _after_ Figure 1 on page 3) but I would suggest leaving this for now given that we may need to make changes to the summary at review time.
