### 2019-05-16 #1449 @zbeekman
Sure! Just ping me with next steps once you’re ready. This will be my first
JOSS review. I’ve read through most of the docs so far. I’ve taken a
cursory look at the repository and will read through the paper.

On Thu, May 16, 2019 at 12:33 PM Arfon Smith <notifications@github.com>
wrote:

> @afron <https://github.com/afron>: I'd be happy to review this submission!
>
> Excellent, thanks @zbeekman <https://github.com/zbeekman>! I'd like to
> find at least two reviewers for this submission so for now please sit tight
> while I find another reviewer.
>
