### 2019-05-17 e1bb1fd
CI: retry downloads to avoid breaks due to network issues
