### 2019-05-18 ad43b74
use new HOMEBREW_CURL_RETRIES var
