### 2019-05-21 #1449 @arfon
@andreas-h @bjarketol - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

The software under consideration is here: https://github.com/WRF-CMake/WRF and you can learn more about the project in this [short overview paper](https://github.com/openjournals/joss-papers/blob/joss.01449/joss.01449/10.21105.joss.01449.pdf)
