### 2019-05-21 #1449 @andreas-h
Yes, I'd be happy to review this.

Tue May 21 13:08:56 GMT+02:00 2019 Arfon Smith <notifications@github.com>:

>
> @andreas-h [https://github.com/andreas-h]   @bjarketol [https://github.com/bjarketol]  - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:  https://joss.readthedocs.io/en/latest/review_criteria.html [https://joss.readthedocs.io/en/latest/review_criteria.html]
>
> The software under consideration is here:  https://github.com/WRF-CMake/WRF [https://github.com/WRF-CMake/WRF]  and you can learn more about the project in this  short overview paper [https://github.com/openjournals/joss-papers/blob/joss.01449/joss.01449/10.21105.joss.01449.pdf]
>
