### 2019-05-21 #1468 @whedon
[REVIEW]: WRF-CMake: integrating CMake support into the Advanced Research WRF (ARW) modelling system
**Submitting author:** @letmaik (<a href="http://orcid.org/0000-0003-3299-1382">Maik Riechert</a>)
**Repository:** <a href="https://github.com/WRF-CMake/WRF" target ="_blank">https://github.com/WRF-CMake/WRF</a>
**Version:** WRF-CMake-4.0.3

- [x] **Version:** WRF-CMake-4.0.3

- [x] **Version:** WRF-CMake-4.0.3
