### 2019-05-21 #1468 @zbeekman
@letmaik Is D. Meyer on GitHub? Trying to ensure no conflicts of interest but, it's hard to guarantee without being able to look up the other author. The institution/association ("Independent Scholar") is not helpful for this.
