### 2019-05-21 #1468 @zbeekman
@arfon @letmaik the version [is listed simply as `4`](https://github.com/openjournals/joss-reviews/issues/1468#issue-446786977) but in the repository I see:

1. [Git tag `v4.0`](https://github.com/WRF-CMake/WRF/releases/tag/v4.0)
2. pre-release with assets, notes & tag `WRF-CMake-4.0`](https://github.com/WRF-CMake/WRF/releases/tag/WRF-CMake-4.0)
3. [pre-release with assets, notes & tag `WRF-CMake-4.0.3`](https://github.com/WRF-CMake/WRF/releases/tag/WRF-CMake-4.0.3)
4. Numerous other tags with `v4.x.y` in them.

- @letmaik: Which one of these should we be reviewing to correspond to the paper?
- @arfon:
  - Do we need this issue to be more precise about the version under review?
  - Is it OK to be reviewing something tagged as a `pre-release`?
