### 2019-05-21 #1468 @arfon
> Do we need this issue to be more precise about the version under review?

It's important that the authors specify which version they're expecting you to review.

> Is it OK to be reviewing something tagged as a `pre-release`?

Yes, as long as the authors are happy to tag a new release at the end of the review based on the `pre-release` content (this often happens for JOSS reviews).
