### 2019-05-21 #1468 @zbeekman
I'm working under the assumption that [pre-release with assets, notes & tag `WRF-CMake-4.0.3`](https://github.com/WRF-CMake/WRF/releases/tag/WRF-CMake-4.0.3) is the version that the authors want reviewed and that will form the basis of the new release.

@letmaik (and/or @dmey): please confirm.
