### 2019-05-22 #1468 @letmaik
Clarification regarding the `pre-release` marker: This only refers to the pre-built binaries under assets which we'd like to declare as "experimental" (meaning use at your own risk, despite our testing efforts), the software itself is not pre-release. Please let us know if you have an idea to make this more clear.

Please review the WRF-CMake-4.0.3 tag. The main difference to the default branch `wrf-cmake` is that we recently added a new CI setup (based on Azure Pipelines) which also produced the plots in the paper; the software is unchanged. See e.g. https://dev.azure.com/WRF-CMake/WRF/_build/results?buildId=254. If you click on Artifacts -> wats_plots, you can find the paper plots at:
A.0 -> d02/T0/OSALL/ext_boxplot.png
A.60 -> d02/T6/OSALL/ext_boxplot.png
B.0 ->  d02/T0/OSALL/nrmse.png
B.60 ->  d02/T6/OSALL/nrmse.png
