### 2019-05-22 #1468 @zbeekman
Should the version be updated at the top of this issue, @arfon?
