### 2019-05-22 #1468 @arfon
> Should the version be updated at the top of this issue, @arfon?

We can do that yes. Can you confirm the version should be `WRF-CMake-4.0.3` @letmaik?
