### 2019-05-22 #1468 @letmaik
@arfon Yes, it's WRF-CMake-4.0.3. How is the process if we need to make changes, e.g. to the docs? Do we have to release a new version which this paper is then ultimately based on? The paper itself is also in a branch currently and not part of 4.0.3.
