+++ .ci/azure-pipelines/matrix.yml
@@ -1,13 +1,15 @@
 parameters:
   templates:
-    Linux: unix.yml
+    Ubuntu: unix.yml
+    CentOS: unix.yml
     macOS: unix.yml
     Windows: windows.yml
   
-  Linux:
-    #CC: gcc-8
+  Ubuntu:
+    CC: gcc-8
+    FC: gfortran-8
+  CentOS:
     CC: gcc
-    #FC: gfortran-8
     FC: gfortran
   macOS:
     CC: gcc-8
@@ -27,10 +29,9 @@ parameters:
   
   # Loop/matrix parameters.
   # Note that these may be overridden by the including yml (typically azure-pipelines.yml).
-  #OS_NAMES: [Windows, Linux, macOS]
-  OS_NAMES: [Linux]
+  #OS_NAMES: [Windows, Ubuntu, macOS]
+  OS_NAMES: [Ubuntu, CentOS]
   BUILD_SYSTEMS: [Make, CMake]
-  #BUILD_SYSTEMS: [CMake]
   #BUILD_TYPES: [Debug, Release]
   BUILD_TYPES: [Debug]
   #MODES: [serial, smpar, dmpar, dm_sm]
@@ -59,9 +60,9 @@ jobs:
       - ${{ if eq(parameters.WATS_DIFF, 'true') }}:
         - template: wats_diff.yml
           parameters:
-            # For Windows, since there is no Make-based reference, compare against Linux.
+            # For Windows, since there is no Make-based reference, compare against Ubuntu.
             ${{ if eq(OS_NAME, 'Windows') }}:
-              OS_NAME_LEFT: Linux
+              OS_NAME_LEFT: Ubuntu
             ${{ if ne(OS_NAME, 'Windows') }}:
               OS_NAME_LEFT: ${{ OS_NAME }}
             OS_NAME_RIGHT: ${{ OS_NAME }}
@@ -77,7 +78,7 @@ jobs:
     - ${{ each MODE in parameters.MODES }}:
       - template: wats_diff.yml
         parameters:
-          OS_NAME_LEFT: Linux
+          OS_NAME_LEFT: Ubuntu
           OS_NAME_RIGHT: macOS
           BUILD_SYSTEM_LEFT: Make
           BUILD_SYSTEM_RIGHT: Make
