+++ .ci/azure-pipelines/unix.yml
@@ -5,23 +5,26 @@ jobs:
   timeoutInMinutes: 0
 
   pool:
-    ${{ if eq(parameters.OS_NAME, 'Linux') }}:
+    ${{ if or(eq(parameters.OS_NAME, 'Ubuntu'), eq(parameters.OS_NAME, 'CentOS')) }}:
       vmImage: ubuntu-16.04
     ${{ if eq(parameters.OS_NAME, 'macOS') }}:
       vmImage: macos-10.13
 
-  container:
-    image: quay.io/pypa/manylinux2010_x86_64:latest
-    options: "--name ci-container -v /usr/bin/docker:/tmp/docker:ro"
+  ${{ if eq(parameters.OS_NAME, 'CentOS') }}:
+    container:
+      image: quay.io/pypa/manylinux2010_x86_64:latest
+      options: "--name ci-container -v /usr/bin/docker:/tmp/docker:ro"
 
   variables:
     ${{ insert }}: ${{ parameters }}
 
   steps:
-  - script: |
-      /tmp/docker exec -t -u 0 ci-container \
-        sh -c "yum install -y sudo"
-    displayName: Set up sudo
+  - ${{ if eq(parameters.OS_NAME, 'CentOS') }}:
+    # https://github.com/Microsoft/azure-pipelines-agent/issues/2043#issuecomment-489692810
+    - script: |
+        /tmp/docker exec -t -u 0 ci-container \
+          sh -c "yum install -y sudo"
+      displayName: Set up sudo
 
   - script: .ci/unix/dump-vm-specs.sh
     displayName: Dump VM specs
