### 2019-05-22 4e627a0
install conda in docker
