+++ .ci/unix/use-conda.sh
@@ -4,6 +4,12 @@ if [ "$(uname)" == "Darwin" ]; then
     echo "##vso[task.prependpath]$CONDA/bin"
     sudo chown -R $USER $CONDA
 elif [ "$(uname)" == "Linux" ]; then
+    if [ "$(lsb_release -i -s)" == "CentOS" ]; then
+        curl -sL https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -o miniconda.sh
+        chmod +x miniconda.sh
+        sudo ./miniconda.sh -b -p /usr/share/miniconda
+        rm miniconda.sh
+    fi
     echo "##vso[task.prependpath]/usr/share/miniconda/bin"
     sudo chown -R $USER /usr/share/miniconda
 else
