+++ azure-pipelines-release.yml
@@ -1,9 +1,9 @@
 # First, create a tag for the new version.
-# Then, run this pipeline on the tag (will not deploy if untagged)
-# to store binaries in a draft GitHub release.
+# Then, run this pipeline on the tag to store binaries in a draft GitHub release.
 
 # Note that these values are also overridden to "none" in Azure DevOps's UI to
 # prevent others to run this pipeline by changing the triggers below and opening a PR.
+# It still wouldn't cause harm as secrets (GitHub credentials) are not available to PRs from forks.
 trigger: none
 pr: none
 
