### 2019-05-23 #1468 @arfon
> @arfon Yes, it's WRF-CMake-4.0.3. How is the process if we need to make changes, e.g. to the docs? Do we have to release a new version which this paper is then ultimately based on? The paper itself is also in a branch currently and not part of 4.0.3.

We ask that authors make a new release that includes the changes made during the review, that way we can be sure the archived version reflects the work of our reviewers too. The paper doesn't need to be part of the final version.
