### 2019-05-26 #1468 @letmaik
@arfon Makes sense. Our project strictly follows the WRF version numbers to avoid confusion. The next version therefore would be WRF-CMake 4.1. For the purpose of reviewing, there is no difference between WRF-CMake 4.0.3 and 4.1 (except the addition of [continuous integration](https://github.com/WRF-CMake/WRF/tree/wrf-cmake/.ci) via [Azure Pipelines](https://dev.azure.com/WRF-CMake/WRF/_build/results?buildId=380&view=logs)). We re-ran the scripts that produce the paper plots and the conclusions are unchanged. We will update the plots as there are minor changes but this doesn't affect the review. To clarify, all feedback from the JOSS review is incorporated into the `wrf-cmake` (our "master") branch from which we will then cut the 4.1 release that the paper is linked to. Does that all make sense?
