### 2019-06-10 59fa5bb
restructure readme [wip] [skip-ci]
