+++ README.md
@@ -1,9 +1,26 @@
 # WRF-CMake [![Build status Azure Pipelines](https://dev.azure.com/WRF-CMake/WRF/_apis/build/status/WRF%20(full)?branchName=wrf-cmake)](https://dev.azure.com/WRF-CMake/WRF/_build/latest?definitionId=5&branchName=wrf-cmake) [![Build status Appveyor](https://ci.appveyor.com/api/projects/status/86508wximkvmf95g/branch/wrf-cmake?svg=true)](https://ci.appveyor.com/project/WRF-CMake/wrf/branch/wrf-cmake) [![Build status Travis CI](https://travis-ci.com/WRF-CMake/WRF.svg?branch=wrf-cmake)](https://travis-ci.com/WRF-CMake/WRF)
 
+  - [Project overview](#project-overview)
+    - [Currently supported platforms](#currently-supported-platforms)
+    - [Currently unsupported features](#currently-unsupported-features)
+  - [Installation](#installation)
+    - [Manually from source](#manually-from-source)
+    - [Using Homebrew or Linuxbrew](#using-homebrew-or-linuxbrew)
+    - [Binary distribution (Experimental)](#binary-distribution-experimental)
+  - [Documentation](#documentation)
+  - [Example usage](#example-usage)
+  - [How to cite](#how-to-cite)
+  - [Testing framework](#testing-framework)
+  - [Changes to be upstreamed](#changes-to-be-upstreamed)
+  - [Copyright and license](#copyright-and-license)
+
+## Project overview
+
 WRF-CMake adds CMake support to the latest version of the [Advanced Research Weather Research and Forecasting](https://www.mmm.ucar.edu/weather-research-and-forecasting-model) model (here WRF, for short) with the intention of streamlining and simplifying its configuration and build process. In our view, the use of CMake provides model developers, code maintainers, and end-users with several advantages such as robust incremental rebuilds, flexible library dependency discovery, native tool-chains for Windows, macOS, and Linux with minimal external dependencies, thus increasing portability, and automatic generation of project files for different platforms.
 
 WRF-CMake is designed to work alongside the current releases of WRF, therefore you can still compile your code using the legacy Makefiles included in WRF and WPS for any of the currently unsupported features.
 
+For more details, please see the short summary paper [WRF-CMake: integrating CMake support into the Advanced Research WRF (ARW) modelling system](https://joss.theoj.org/papers/9a87d84b2ed00ed82a6e297a4c34b3cf) on [The Journal of Open Source Software](https://joss.theoj.org/) website.
 
 ### Currently supported platforms
 
@@ -27,7 +44,21 @@ WRF-CMake is designed to work alongside the current releases of WRF, therefore y
 
 ## Installation
 
-The installation of WRF-CMake is straightforward thanks to the downloadable pre-built binaries for most Linux distributions (specifically [ RPM-based and Debian-based distribution-compatible](https://en.wikipedia.org/wiki/List_of_Linux_distributions)), macOS, and Windows (see [binary distribution](#binary-distribution-experimental) below) -- most users wishing to run WRF on their system can simply download the pre-compiled binaries without the need of building from source. Alternately, to build WRF from source, please refer to the [source distribution](#source-distribution) section below. HPC users, or users seeking to run WRF in the 'most optimal' configuration for their system, are advised to build WRF-CMake from source.
[…]
