+++ .ci/azure-pipelines/matrix.yml
@@ -8,11 +8,11 @@ parameters:
     Windows: windows.yml
   
   Linux:
-    CC: gcc-8
-    FC: gfortran-8
+    CC: gcc-9
+    FC: gfortran-9
   macOS:
-    CC: gcc-8
-    FC: gfortran-8
+    CC: gcc-9
+    FC: gfortran-9
   Windows:
     CC: gcc
     FC: gfortran
