### 2019-06-10 1264222
Add community guidelines/contributing. Fixes #23.
