+++ CONTRIBUTING.md
@@ -1,0 +1,32 @@
+# How to contribute
+
+Thank you for considering contributing to the WRF-CMake project. Please note that the aim of WRF-CMake is to bring CMake support to WRF and WPS therefore, any contributions made to this project should only consist of changes/fixes to issues related to building WRF or WPS with CMake support, pre-built binaries, Homebrew/Linuxbrew support, and Continuous Integration (CI) and testing.
+
+For any issues/contributions related with the WRF or WPS model, please see the [WRF](https://github.com/wrf-model/WRF) and [WPS](https://github.com/wrf-model/WPS) repositories directly. For any question/issues with the model physics please refer to the [WRF/WRFDA/WRF-Chem/MPAS Users' Support Forum](http://forum.mmm.ucar.edu/phpBB3/).
+
+In general, you can contribute by reporting an issue or by directly contributing to the source code. For the latter, fork our repository, clone the fork, make your changes, and create a pull request (PR) with a **clear description** of your changes -- if you are unfamiliar about forking/creating PR, please see [this guide](https://guides.github.com/activities/forking/) first. If/when your changes are merged, you will appear as a [Contributors](https://github.com/WRF-CMake/WRF/graphs/contributors). For specific instructions on how to report a bug or how to submit a feature request, please see below:
+
+- [Report a bug](#report-a-bug)
+- [Request a new feature](#request-a-new-feature)
+- [Coding conventions](#coding-conventions)
+- [Versioning](#versioning)
+
+
+If you would like to know more about specific development guidelines, please refer to our [development notes](DEVELOP.md).
+
+## Report a bug
+
+Before creating bug reports, please check if similar issue have already been reported [here](https://github.com/WRF-CMake/WRF/issues). If none exist please create a new issue and include as many details as possible using the required template.
+
+## Request a new feature
+
+Before creating a feature request, please check if a similar feature has been already asked [here](https://github.com/WRF-CMake/WRF/issues). If none exist please create a new feature request and include as many details as possible using the required template.
+
+## Coding conventions
+
+Same as in the [WRF](https://github.com/wrf-model/WRF) and [WPS](https://github.com/wrf-model/WPS).
+
+
+## Versioning
+
+WRF-CMake and WPS-CMake fully tracks releases made by the [WRF](https://github.com/wrf-model/WRF) and [WPS](https://github.com/wrf-model/WPS) model.
