### 2019-06-11 618c8a9
add _OPENMP preprocessor def if OpenMP is enabled
