### 2019-06-11 4e09c69
minor fixes to bug_reports and readme
