+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -22,14 +22,8 @@ If applicable, add screenshots to help explain your problem.
 
 **Desktop (please complete the following information):**
  - OS: [e.g. iOS]
- - Browser [e.g. chrome, safari]
- - Version [e.g. 22]
-
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
+ - Compiler [e.g. gcc/gfortran, icc/ifortran]
+ - Version [e.g. 8.1.0]
 
 **Additional context**
 Add any other context about the problem here.
