+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -21,8 +21,8 @@ A clear and concise description of what you expected to happen.
 If applicable, add screenshots to help explain your problem.
 
 **Desktop (please complete the following information):**
- - OS: [e.g. iOS]
- - Compiler [e.g. gcc/gfortran, icc/ifortran]
+ - OS: [e.g. Ubuntu, macOS, Windows]
+ - Compiler [e.g. gcc/gfortran, icc/ifort]
  - Version [e.g. 8.1.0]
 
 **Additional context**
