### 2019-06-11 3d14955
calrify sys info section in bug_report.md
