### 2019-06-11 bca3fed
Add community guidelines/contributing etc. Closes #23. (#30)
Adds community guidelines/contributing etc. as requested in #23.
