### 2019-06-11 4c8175b
revert setting _OPENMP, automatically done by the compiler
