### 2019-06-11 42f3a0f
travis: add ubuntu 16.04 builds with gcc 9 & switch from dmpar to dm_sm to exercise both MPI and OpenMP
