+++ .travis.yml
@@ -28,7 +28,29 @@ jobs:
         - BUILD_TYPE=Debug
         - CC=gcc
         - FC=gfortran
-        - MODE=dmpar
+        - MODE=dm_sm
+
+    - stage: build
+      os: linux
+      dist: xenial
+      sudo: required
+      env:
+        - BUILD_SYSTEM=Make
+        - BUILD_TYPE=Debug
+        - CC=gcc-9
+        - FC=gfortran-9
+        - MODE=serial
+
+    - stage: build
+      os: linux
+      dist: xenial
+      sudo: required
+      env:
+        - BUILD_SYSTEM=Make
+        - BUILD_TYPE=Debug
+        - CC=gcc-9
+        - FC=gfortran-9
+        - MODE=dm_sm
     
     - stage: build
       os: osx
@@ -50,7 +72,7 @@ jobs:
         - BUILD_TYPE=Debug
         - CC=gcc-9 # GNU via Homebrew
         - FC=gfortran-9
-        - MODE=dmpar
+        - MODE=dm_sm
 
     # CMake builds
 
@@ -83,6 +105,34 @@ jobs:
         - NESTING=basic
 
     - stage: build
+      os: linux
+      dist: xenial
+      sudo: required # Need sudo to install netcdf via apt as not whitelisted 
+      env:
+        - BUILD_SYSTEM=CMake
+        - BUILD_TYPE=Debug
+        - CC=gcc-9
+        - FC=gfortran-9
+        - GRIB1=0
+        - GRIB2=1
+        - MODE=serial
+        - NESTING=basic
+
+    - stage: build
+      os: linux
+      dist: xenial
+      sudo: required
+      env:
+        - BUILD_SYSTEM=CMake
+        - BUILD_TYPE=Debug
+        - CC=gcc-9
+        - FC=gfortran-9
+        - GRIB1=1
+        - GRIB2=0
+        - MODE=dm_sm
+        - NESTING=basic
+
+    - stage: build
       os: osx
       osx_image: xcode9.2 # https://github.com/travis-ci/travis-ci/issues/9640
       language: generic
@@ -107,7 +157,7 @@ jobs:
         - FC=gfortran-9
         - GRIB1=1
         - GRIB2=1
-        - MODE=dmpar
+        - MODE=dm_sm
         - NESTING=basic
 
 before_install:
