+++ .appveyor.yml
@@ -21,7 +21,7 @@ environment:
     - BUILD_TYPE: Debug
       CC: gcc
       FC: gfortran
-      MODE: dmpar
+      MODE: dm_sm
       NESTING: basic
       GRIB1: 0
       GRIB2: 1
