### 2019-06-12 cb63872
minor readme fixes [skip ci]
