### 2019-06-12 ef7032a
minor edits to CONTRIBUTING.md [skip ci]
