### 2019-06-12 446a550
switch netcdf-fortran download to http for robustness & remove incorrect comment
