+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @techreport{Skamarock2008,
 title = {{A Description of the Advanced Research WRF Version 3}},
-author = {Skamarock, W.C. and Klemp, J.B. and Dudhi, J. and Gill, D.O. and Barker, D.M. and Duda, M.G. and Huang, X.-Y. and Wang, W. and Powers, J.G.},
+author = {Skamarock, William C. and Klemp, J.B. and Dudhi, J. and Gill, D.O. and Barker, D.M. and Duda, M.G. and Huang, X.-Y. and Wang, W. and Powers, J.G.},
 institution = {{NCAR} Technical Note},
 month = {June},
 year = {2008},
