### 2019-06-30 #1468 @andreas-h
First of all, let me express my apologies for being so late with this review.

Regarding the installation instructions, in https://github.com/WRF-CMake/WRF/blob/wrf-cmake/doc/cmake/INSTALL.md, I believe the authors should clearly state how to use the CMake system in typical situations occuring in a WRF workflow, e.g., what steps to do after a modification of the Registry.  In upstream WRF, I can use the `./clean` and `./compile` scripts, but with WRF-CMake, I wouldn't know what to do. This is confusing for users who are alrady familiar with WRF.

Maybe a comparison table "upstream WRF vs. WRF-CMake" would be useful.
