### 2019-06-30 #1468 @andreas-h
It's great that we can now build WRF on Windows without hassle, and that pre-built binaries can be distributed.  That's awesome!

However, I do not understand why the authors chose the way to fork the upstream repository.  In my mind, this creates major issues regarding sustainability of WRF-CMake on one hand, and regarding traceability on the other hand.

If CMake support were included in upstream WRF, as an option, this would without doubt be ideal.

I can understand that there might be reasons to not go this way.  However, offering a fork of the upstream repository is bad for traceability.  For example, the authors say that the https://github.com/WRF-CMake/WRF/tree/wrf-cmake branch is supposed to be WRF version 4.0.3.  However, I'm not sure how the user is supposed to verify this.  When I look at https://github.com/wrf-model/WRF/compare/release-v4.0.3...WRF-CMake:wrf-cmake there's almost 200 commits difference between the two suppoedly equivalent branches, many of them by upstream WRF authors that seem not to be related to CMake integration at all.  So how can we be sure that this is indeed WRF-4.0.3?

For these reasons, if including in upstream WRF was not possible (did the authors even try?) my personal approach would have been to create a repository for a scripted addition of CMake support to WRF, including all the needed patches and so on.  Of course, this could be built on diff'ing between branches of upstream WRF and a fork.

This way, the user would be able to understand what is the difference between WRF-CMake and upstream WRF better, the relationship between the two projects would be a lot more transparent. As it stands now, what do I do if there's a new upstream WRF release for which the WRF-CMake authors have not released yet?  Can I just apply the upstream commits to the WRF-CMake repository?

I believe the manuscript and repository should address these questions.
