### 2019-06-30 #1468 @andreas-h
Another issue: Given that WRF-Chem is included in the upstream WRF source tree since version 4, the authors should make explicit if WRF-Chem is also supported in WRF-CMake and how to compile "WRF-CMake-Chem" with and without KPP support.
