### 2019-06-30 f89fdef
debug output
