+++ .ci/unix/delocate.sh
@@ -10,7 +10,9 @@ cd $SCRIPTDIR/../..
 if [ "$(uname)" == "Darwin" ]; then
 
     pip3 install delocate
+    delocate-listdeps --all --depending build/install/main
     delocate-path build/install/main
+    delocate-listdeps --all --depending build/install/main
 
 else
     echo "Unknown OS: $(uname)"
