### 2019-07-03 67d3116
make ci script for dependency install more generic allows easier use in docker containers
