+++ .ci/azure-pipelines/matrix.yml
@@ -11,8 +11,8 @@ parameters:
     CC: gcc-8
     FC: gfortran-8
   macOS:
-    CC: gcc-8
-    FC: gfortran-8
+    CC: gcc-9
+    FC: gfortran-9
   Windows:
     CC: gcc
     FC: gfortran
