### 2019-07-03 9d052fa
install lsb-release missing in docker
