### 2019-07-03 a27bec7
install lsb-release only when needed
