+++ .ci/unix/dump-vm-specs.sh
@@ -14,7 +14,9 @@ if [ "$(uname)" == "Darwin" ]; then
     sudo scutil --get HostName || true
     sudo scutil --get LocalHostName || true
 elif [ "$(uname)" == "Linux" ]; then
-    sudo apt install -y lsb-release
+    if [ "$(which lsb_release)" == "" ]; then
+        sudo apt install -y lsb-release
+    fi
     lsb_release -a
     free -m
     lscpu
