### 2019-07-04 cbb795c
install miniconda if missing (for docker)
