+++ .ci/unix/use-conda.sh
@@ -7,8 +7,14 @@ if [ "$(uname)" == "Darwin" ]; then
     echo "##vso[task.prependpath]$CONDA/bin"
     sudo chown -R $USER $CONDA
 elif [ "$(uname)" == "Linux" ]; then
+    if [ ! -d /usr/share/miniconda ]; then
+        curl -sL https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -o miniconda.sh
+        chmod +x miniconda.sh
+        sudo ./miniconda.sh -b -p /usr/share/miniconda
+        rm miniconda.sh
+    fi
     echo "##vso[task.prependpath]/usr/share/miniconda/bin"
-    sudo chown -R $USER /usr/share/miniconda
+    sudo chown -R $(id -u -n) /usr/share/miniconda
 else
     echo "##vso[task.prependpath]$CONDA\Scripts"
     # We don't use conda activate as we would have to repeat that in each task,
