### 2019-07-04 9d3a715
use curl instead of git for downloading wats avoids installing git inside container jobs
