### 2019-07-04 add2fa0
cleanup curl usage
