### 2019-07-04 b3f0d3d
avoid powershell as it may be missing in container
