+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -14,4 +14,4 @@ A clear and concise description of what you want to happen.
 A clear and concise description of any alternative solutions or features you've considered.
 
 **Additional context**
-Add any other context or screenshots about the feature request he
+Add any other context or screenshots about the feature to be requested.
