### 2019-07-06 a900b46
remove unused link [skip ci]
