+++ paper/paper.bib
@@ -1,12 +1,11 @@
-@techreport{Skamarock2008,
-title = {{A Description of the Advanced Research WRF Version 3}},
-author = {Skamarock, William C. and Klemp, J.B. and Dudhi, J. and Gill, D.O. and Barker, D.M. and Duda, M.G. and Huang, X.-Y. and Wang, W. and Powers, J.G.},
-institution = {{NCAR} Technical Note},
-month = {June},
-year = {2008},
-doi = {10.5065/D6DZ069T},
-isbn = {NCAR/TN–468+STR},
-pages = {113}
+@article{Skamarock2019,
+  title = {A Description of the Advanced Research WRF Model Version 4},
+  author = {Skamarock,  William C. and Klemp,  Joseph B. and Dudhia,  Jimy and Gill,  David O. and Liu,  Zhiquan and Berner,  Judith and Wang,  Wei and Powers,  Jordan G. and Duda,  Michael G. and Barker,  Dale M. and Huang,  Xiang-Yu},
+  publisher = {UCAR/NCAR},
+  year = {2019},
+  doi = {10.5065/1dfh-6p97},
+  isbn = {NCAR/TN-556+STR},
+  pages = {145}
 }
 
 @article{Powers2017,
