+++ paper/paper.bib
@@ -1,6 +1,7 @@
-@article{Skamarock2019,
+@techreport{Skamarock2019,
   title = {A Description of the Advanced Research WRF Model Version 4},
   author = {Skamarock,  William C. and Klemp,  Joseph B. and Dudhia,  Jimy and Gill,  David O. and Liu,  Zhiquan and Berner,  Judith and Wang,  Wei and Powers,  Jordan G. and Duda,  Michael G. and Barker,  Dale M. and Huang,  Xiang-Yu},
+  institution = {{NCAR Technical Note NCAR/TN-556+STR}},
   publisher = {UCAR/NCAR},
   year = {2019},
   doi = {10.5065/1dfh-6p97},
