### 2019-07-15 #1468 @zbeekman
I have concluded my review and enthusiastically recommend this paper for publication, under the one condition that https://github.com/WRF-CMake/wrf/pull/29 is eventually merged to add contributing guidelines and clarify issue/bug reporting etc.

Furthermore, I would recommend that the authors and editors examine the second paragraph of the paper; the sentence construction is a bit circuitous and hard to follow at times. Breaking up longer sentences, especially those with lists or enumerations would help make that paragraph easier to read and comprehend.

Lastly, the paper ends somewhat suddenly; concluding remarks and further exploration of why the most impactful change appears to be choice of OS, even if only speculative, would be interesting.

Furthermore, I deeply hope that the WRF package maintainers decide to integrate these changes upstream, as they will surely increase the ease of use of WRF for new users, if not also increase the likelihood that WRF may eventually be packaged and distributed via package managers such as Debian's `apt` or Mac Homebrew. The contributed CMake build system provides a degree of normalization with software development best practices, and should provide for increased portability and flexibility while maintaining or decreasing build system complexity and dependencies. 💯
