### 2019-08-12 #1468 @andreas-h
there's a typo in the install docs, see WRF-CMake/wrf/pull/37/files
