### 2019-08-12 #1468 @andreas-h
> @andreas-h I'll address your easy/technical questions first before I talk about the bigger issue you raised.
>
> > Maybe a comparison table "upstream WRF vs. WRF-CMake" would be useful.
>
> Great suggestion! This is exactly the kind of feedback we need, since we ourselves are more or less blind to such issues now. I've updated the page in [WRF-CMake/wrf@cc030c8](https://github.com/WRF-CMake/wrf/commit/cc030c8a1809e4c730cf6766be7d5f2f3013ca1d). Let us know if there's something else that would be good to include.

Thanks, that new version of the documentation is a lot more helpful to users who already "know" WRF.

> > When I look at wrf-model/WRF@release-v4.0.3...WRF-CMake:wrf-cmake there's almost 200 commits difference between the two suppoedly equivalent branches, many of them by upstream WRF authors that seem not to be related to CMake integration at all. So how can we be sure that this is indeed WRF-4.0.3?
>
> Also a great comment. You accidently compared our "master" which is already at 4.1 against upstream 4.0.3. That's why you saw so many changes. I've updated the release notes to include a new section on how our release relates to upstream WRF, e.g.: https://github.com/WRF-CMake/WRF/releases/tag/WRF-CMake-4.0.3. This should make traceability a non-issue.

Agreed.  The way you describe the relation in the release notes is very helpful, especially with the link.  It would be great to have that in the release notes always.

> > Given that WRF-Chem is included in the upstream WRF source tree since version 4, the authors should make explicit if WRF-Chem is also supported in WRF-CMake and how to compile "WRF-CMake-Chem" with and without KPP support.
>
> See the [unsupported features](https://github.com/WRF-CMake/WRF/tree/WRF-CMake-4.0.3#currently-unsupported-features) section in the README.

Thanks, that was me not reading properly.

> Regarding your suggestion of creating a repository for a "scripted addition of CMake", including patch files: We decided against that because we still have upstreaming on the radar, and we want to keep our fork PRable, meaning in a state where the WRF team can just take it, without rewriting lots of plumbing that would result from an alternative scripted approach. Since currently every new release of WRF has to be downstreamed to WRF-CMake and can't be done automatically, we don't see an advantage for a scripted approach. For us it's much easier to work with since we only deal with one repository. For any experiments (e.g. research on some schemes), we just fork WRF-CMake (or create a branch) and we can hack in WRF files as well as CMake files as we wish, which would be much harder if we would have to use patch files for WRF etc. or just the fact of dealing with two repositories (actual WRF, and the CMake additions).

Agreed.  Having your fork PRable is indeed very helpful.
