### 2019-08-12 WRF-CMake/WRF#38 @andreas-h
Add information about re-compilation to install instructions
When working with WRF, sometimes one needs to re-compile the code (e.g., after changing some things in the source code).  When changing the Registry, even a full re-compile is necessary (e.g., `./clean -a && ./compile em_real && ./compile wrf`).

From the compilation instructions here, it is unclear to me how this relates to the WRF-CMAKE build system.

Do I always follow the `cmake ... && make install` ?  Or do I have to clean somehow when changing the registry?

This should be added to the documentation.
