### 2019-08-12 #1468 @andreas-h
Some more info regarding cleaning / re-compiling when modifying the source code should be added to the compilation instructions, see WRF-CMake/wrf/issues/38
