### 2019-08-12 #1468 @andreas-h
The paper is well-written.  I agree with @zbeekman that it ends rather abruptly.  Also, I think that including the comparison *howto-do-A-in-upstream-WRF* vs *howto-do-A-in-WRF-CMAKE*I (as discussed above and in WRF-CMake/wrf/issues/38) should be included in the paper.

I am a bit shocked by the NRMSE plot.  Do I read this right, that after 60 minutes simulation time, `w` differs between WRF-CMAKE and upstream WRF by ~30%? That's enormous, and I personally don't see how this large error could be acceptable.

I think that the metric the authors use is not the one relevant for this paper.  In order to see that WRF-CMAKE is not different from upstream WRF, it would be more interesting to see the $\delta$ and NRMSE between upstream WRF and WRF-CMAKE for each compiler/platform/optimization setting individually.  What I mean is, e.g., for MacOS/Debug/dmpar, how different is WRF-CMAKE from upstream WRF?  By comparing everything to upstream WRF/Debug/serial, it is very hard for the reader to distinguish between the impact of the platform/compiler/optimization on the one hand and the WRF-CMAKE on the other hand.  (What I mean is that in fact, most of the horrendous differences are not related to WRF-CMAKE but to the other factors, which makes WRF-CMAKE look unnecessarily bad in this plot).

As for the test suite, I would appreciate some more information:  There is the *wats* repository on Github. I believe this was used to create Fig. 1 somehow? It would be good if the repository would include information about how the tests leading to Fig. 1 were conducted, e.g., by pointing to the namelist(s) in the wats repository. As it stands, the claims made about the $\delta$ and NRMSE differences are not reproducible by the reader.
