### 2019-08-12 #1468 @andreas-h
Having said all this, I think WRF-CMAKE is a valuable contribution.  If it had been around when I started working with WRF, it had saved me several days of work (and frustration).

There are still three boxes I haven't checked above:
1. *Installation* I'll check off once a test build I'm doing on our system completes successfully
2. *Example usage* I'll check off once the authors have addressed WRF-CMake/wrf/issues/38
3. *Automated tests* I'll check off once the authors have addressed my comment about the test suite above.

Lastly, I want to apologize for being so painfully late.  It's been a crazy summer so far, and I am really sorry for being so unreliable!

I really hope that your changes make it upstream at some point, because they're very helpful!
