### 2019-08-12 WRF-CMake/WRF#38 @letmaik
Please see https://github.com/WRF-CMake/wrf/blob/wrf-cmake/doc/cmake/INSTALL.md#build-and-install-wrf-cmake. There's a point saying:
> - If you change any registry files, then just re-run `make install`.

Is that what you were looking for?
