### 2019-08-12 WRF-CMake/WRF#38 @andreas-h
> Please see https://github.com/WRF-CMake/wrf/blob/wrf-cmake/doc/cmake/INSTALL.md#build-and-install-wrf-cmake [https://github.com/WRF-CMake/wrf/blob/wrf-cmake/doc/cmake/INSTALL.md#build-and-install-wrf-cmake] . There's a point saying:
 >
 >>
 >> * If you change any registry files, then just re-run make install .
 >>
 >>
 >
 > Is that what you were looking for?

Yes that's fine, thank you for pointing out what I was apparently too blind to see :-/
