+++ paper/paper.md
@@ -36,22 +36,21 @@ WRF-CMake is available as a free and open-source project on GitHub at [https://g
 
 # Testing
 
-A fundamental aspect of software development is testing. Ideally, model components should be tested individually and under several testing methodologies [@Feathers2004], however, given that the WRF framework does not offer a way to unit test its components, we instead run build and integration tests to evaluate the effects of our changes.
+A fundamental aspect of software development is testing. Ideally, model components should be tested individually and under several testing methodologies [@Feathers2004]. As the WRF framework does not offer a way to unit test its components, we instead run build and integration tests to evaluate the effects of our changes.
 
-In WRF-CMake, build tests are performed for all variants and supported platforms at every code commit while integration tests are performed by running several simulations with different configurations (Table 1) with a subset of namelists from the official [WRF Testing Framework](https://github.com/wrf-model/WTF) using the [WRF-CMake Automated Testing Suite (WATS)](https://github.com/WRF-CMake/wats) at any major code changes (e.g. before merging pull requests) to constrain the computing resources used for testing.
+Build and integration tests are performed for all supported configurations (Table 1) using continuous integration services. Build tests are run at every code commit. Integration tests are run using the [WRF-CMake Automated Testing Suite (WATS)](https://github.com/WRF-CMake/wats) with a subset of namelists from the official [WRF Testing Framework](https://github.com/wrf-model/WTF) only at major code changes (e.g. before merging pull requests) to constrain computing resources.
+
+| Dimension  | Variant                             |
+| ---------- | ----------------------------------- |
+| OS         | `Linux`, `macOS`, `Windows`         |
+| Build tool | `Make`, `CMake`                     |
+| Build type | `Debug`, `Release`                  |
+| Mode       | `serial`, `dmpar`, `smpar`, `dm_sm` |
 
 
-| Dimension       | Variants                             |
-| --------------- | -------------------------------------|
-| OS              | Linux, macOS, Windows                |
-| Build tool      | Make, CMake                          |
-| Build type      | Debug, Release                       |
-| Mode            | serial, dmpar, smpar, dm_sm          |
+Table: Configurations used for build and integration tests. `Make`: original WRF build system files; `CMake`: this paper; `Debug`: compiler-optimizations disabled; `Release`: compiler-optimizations enabled; `serial`: single processor support; `dmpar`: multiple processor (MP) with distributed memory support (MPI), `smpar` with shared memory support (OpenMP), and `dm_sm` MP with distributed and shared memory support (MPI and OpenMP).
 
-
-Table: Configurations used for running build and integration tests. For build tool, *Make* refers to the original hand-crafted WRF build system files, while *CMake* refers to the work presented in this paper. For build type, *Debug* has compiler optimizations disabled, while *Release* enabled them. For mode, *serial* is single processor support, *dmpar* is multiple processor (MP) with distributed memory support (MPI), smpar is MP with shared memory support (OpenMP), and *dm_sm* is MP with distributed and shared memory support (MPI and OpenMP).
-
[…]
