### 2019-08-14 adae663
more fixes to paper
