+++ paper/paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 # Summary
 
-The Weather Research and Forecasting model (WRF[^1]) model [@Skamarock2019] is an atmospheric modelling system widely used in operational forecasting and atmospheric research [@Powers2017]. WRF is released as a free and open-source software and officially supported to run on Unix and Unix-like operating systems and on different hardware architectures from single-core computers to multi-core supercomputers. Its current build system relies on several bespoke hand-written Makefiles, and Perl and Shell scripts that have been supported and extended during the many years of development.
+The Weather Research and Forecasting model (WRF[^1]) model [@Skamarock2019] is an atmospheric modelling system widely used in operational forecasting and atmospheric research [@Powers2017]. WRF is released as a free and open-source software and officially supported to run on Unix and Unix-like operating systems and on several hardware architectures from single-core computers to multi-core supercomputers. Its current build system relies on several bespoke hand-written Makefiles, and Perl and Shell scripts that have been supported and extended during the many years of development.
 
 The use of build script generation tools, that is, tools that generate files for native build systems from specifications in a high-level language, rather than manually maintaining build scripts for different environments and platforms, can be useful to reduce code duplication and to minimize issues with code not building correctly [@Hoffman2009], to make software more accessible to a broader audience, and the support less expensive [@Heroux2009]. As such, a common build script generation tool is [CMake](https://cmake.org/). Today, CMake is employed in several projects such as [HDF5](https://www.hdfgroup.org/), [EnergyPlus](https://energyplus.net/), and [ParaView](https://www.paraview.org/) to build modern software written in C, C++, and Fortran in high performance computing (HPC) environments and, by CERN, to allow users to easily set-up and build several million lines of C++ and Python code used in the offline software of the ATLAS experiment at the Large Hadron Collider (LHC) [@Elmsheuser2017].
 
@@ -38,7 +38,7 @@ WRF-CMake is available as a free and open-source project on GitHub at [https://g
 
 A fundamental aspect of software development is testing. Ideally, model components should be tested individually and under several testing methodologies [@Feathers2004]. As the WRF framework does not offer a way to unit test its components, we instead run build and integration tests to evaluate the effects of our changes.
 
-Build and integration tests are performed for all supported configurations (Table 1) using continuous integration services. Build tests are run at every code commit. Integration tests are run using the [WRF-CMake Automated Testing Suite (WATS)](https://github.com/WRF-CMake/wats) with a subset of namelists from the official [WRF Testing Framework](https://github.com/wrf-model/WTF) only at major code changes (e.g. before merging pull requests) to constrain computing resources.
+Build and integration tests are performed for all supported configurations (Table 1) using CI services. Build tests are run at every code commit. Integration tests are run using the [WRF-CMake Automated Testing Suite (WATS)](https://github.com/WRF-CMake/wats) with a subset of namelists from the official [WRF Testing Framework](https://github.com/wrf-model/WTF) only at major code changes (e.g. before merging pull requests) to constrain computing resources.
 
 | Dimension  | Variant                             |
 | ---------- | ----------------------------------- |
@@ -48,9 +48,9 @@ Build and integration tests are performed for all supported configurations (Tabl
 | Mode       | `serial`, `dmpar`, `smpar`, `dm_sm` |
 
 
[…]
