### 2019-08-14 5e82760
more paper fixes [skip ci]
