### 2019-08-14 af5c146
paper: minor changes [skip ci]
