+++ paper/paper.md
@@ -48,9 +48,9 @@ Build and integration tests are performed for all supported configurations (Tabl
 | Mode       | `serial`, `dmpar`, `smpar`, `dm_sm` |
 
 
-Table: Configurations used for build and integration tests. `Make`: original WRF build system files; `CMake`: this paper; Compiler-optimization: `Debug` disabled, `Release` enabled; Processor support: `serial` single, `dmpar` multiple with distributed memory (MPI), `smpar` multiple with shared memory (OpenMP), `dm_sm` multiple with MPI and OpenMP.
+Table: Configurations used for build and integration tests. `Make`: original WRF build system files, `CMake`: this paper; `Debug`: compiler optimizations disabled, `Release` enabled; `serial`: single processor, `dmpar`: multiple with distributed memory (MPI), `smpar`: multiple with shared memory (OpenMP), `dm_sm`: multiple with MPI and OpenMP.
 
-As noted by Hodyss and Majumdar [-@Hodyss2007], and Geer [-@Geer2016], the high sensitivity to initial conditions of dynamical systems, such as the ones used in weather models, can lead to large differences in skill between any two forecasts. It is this high sensitivity to initial conditions that can obscure the source of model error, whether this originates from a change in compiler or architecture, an actual coding error, or indeed, the intrinsic nature of the dynamical system employed. As a result, we evaluate the impact of our changes by comparing the outputs from integration tests against a reference build defined as `Linux/Make/Debug/serial` using [relative percentage error](https://en.wikipedia.org/w/index.php?title=Approximation_error&oldid=878331002#Formal_Definition) ($\delta$) and [normalised root mean square error](https://en.wikipedia.org/w/index.php?title=Root-mean-square_deviation&oldid=893196204#Normalized_root-mean-square_deviation) (NRMSE) at the start of the simulation (Figure 1, A0 and B0) and after 60 minutes (Figure 1, A60 and B60). Both $\delta$ and NRMSE are computed per domain for all grid-points on all vertical levels. Normalizing factors are computed per grid-point for $\delta$ and per domain, per quantity, per variant, on all vertical-levels and grid-points for NRMSE. $\boldsymbol{\delta}$ represents the vector of all $\delta$s per domain.
+As noted by Hodyss and Majumdar [-@Hodyss2007], and Geer [-@Geer2016], the high sensitivity to initial conditions of dynamical systems, such as the ones used in weather models, can lead to large differences in skill between any two forecasts. It is this high sensitivity to initial conditions that can obscure the source of model error, whether this originates from a change in compiler or architecture, an actual coding error, or indeed, the intrinsic nature of the dynamical system employed. As a result, we evaluate the impact of our changes by comparing the outputs from integration tests against a reference configuration defined as `Linux/Make/Debug/serial` using [relative percentage error](https://en.wikipedia.org/w/index.php?title=Approximation_error&oldid=878331002#Formal_Definition) ($\delta$) and [normalised root mean square error](https://en.wikipedia.org/w/index.php?title=Root-mean-square_deviation&oldid=893196204#Normalized_root-mean-square_deviation) (NRMSE) at the start of the simulation (Figure 1, A0 and B0) and after 60 minutes (Figure 1, A60 and B60). Both $\delta$ and NRMSE are computed per domain for all grid points on all vertical levels. Normalizing factors are computed per grid point for $\delta$ and per domain, per quantity, per variant, on all vertical levels and grid points for NRMSE. $\boldsymbol{\delta}$ represents the vector of all $\delta$s per domain.
 
 
 | Symbol   | Name                                  | Unit                   |
@@ -65,7 +65,7 @@ As noted by Hodyss and Majumdar [-@Hodyss2007], and Geer [-@Geer2016], the high 
 Table: WRF prognostic variables evaluated during integration tests.
 
 
[…]
