### 2019-08-14 c40cf91
paper: remove duplication & fix layout
