### 2019-08-14 #1468 @letmaik
We took another pass over the paper and also fixed the abrupt end. We opted not to include a comparison table of commands between existing build system and CMake as the paper is not meant to be a manual or tutorial. We'd like to avoid duplication with the online docs and don't find it significant enough to be included in the paper.

From our point of view we're done now. If the paper gets accepted then the final steps would be to merge https://github.com/WRF-CMake/wrf/pull/29, the `joss-paper` branch, and cut a new version.
