### 2019-08-15 #1468 @arfon
@andreas-h - please take another look at this when you get a chance. I'm hoping we can wrap this up soon :-)
