### 2019-08-16 #1468 @andreas-h
> > I am a bit shocked by the NRMSE plot. Do I read this right, that after 60 minutes simulation time, `w` differs between WRF-CMAKE and upstream WRF by ~30%? That's enormous, and I personally don't see how this large error could be acceptable.
>
> No, you misread that. All errors are against the reference Linux/Make/Debug/serial. On the y axis of the plot you see what is compared against. You can see that both Make and CMake generate such high errors, simply due to differences in operating system or build mode (Debug, Release). The paragraph above Table 1 explains why this is the case.

Agreed. I had actually already figured that out but seem to have failed to remove that sentence from my comment.

> > As for the test suite, I would appreciate some more information: There is the _wats_ repository on Github. I believe this was used to create Fig. 1 somehow? It would be good if the repository would include information about how the tests leading to Fig. 1 were conducted, e.g., by pointing to the namelist(s) in the wats repository.
>
> I extended the docs for WATS: https://github.com/WRF-CMake/wats and also added a link to that from the wrf repo here: https://github.com/WRF-CMake/wrf#testing Let us know if that's good enough.

Yes, thanks, that's good enough.

> Regarding the abrupt end and the comparison table, we'll think about this and will respond/fix this in a few days.

I agree with the concluding remarks you added.

> > I think that the metric the authors use is not the one relevant for this paper. In order to see that WRF-CMAKE is not different from upstream WRF, it would be more interesting to see the $\delta$ and NRMSE between upstream WRF and WRF-CMAKE for each compiler/platform/optimization setting individually. What I mean is, e.g., for MacOS/Debug/dmpar, how different is WRF-CMAKE from upstream WRF?
>
> I agree this would seem intuitive and we actually did this initially but it turned out not to have much value. I'll try to explain with an example: WRF-Make does not use compilation optimization options consistently across its Fortran files. Whether that is on purpose or not is not clear. In WRF-CMake we aim to be consistent and only deviate if absolutely necessary. This results in sometimes using different optimzation flags compared to WRF-Make, which subsequently may introduce different rounding errors etc. due to how maths operations are compiled to machine code. Because of this and the nature of the model (sensitivity, see paper) the simulation results in some cases may differ a lot after a certain amount of simulation time. The point is that this is expected. And indeed if we compare WRF-Make with itself under different operating systems or optimization options we see the same effects (see plots).

This is a very nice explanation of the issue.  I think the paper would benefit if you would include this explanation in the paper.

> Therefore, simply comparing CMake vs Make per configuration/system would show an error that doesn't mean much.

Here, I disagree somewhat.  In the scope of this paper, which is to introduce WRF-CMAKE, I think the difference between upstream WRF and WRF-CMAKE is interesting.  It answers the important question *Given that I have been using upstream WRF so far, will I see significant differences if I switch to WRF-CMAKE now with my existing simulation setup*.

> However, comparing all the runs against a single reference shows the variability of WRF under different conditions, and if CMake is within the same ballpark then we can assume we're not doing something dramatically bad. We realize that this is not the ultimate test, but to do better you would have to run massive ensemble simulations and then look at the average and compare them against each other. We don't have the resources to do that, hence we opted for this approach. You can get a good indication of the sanity of our CMake implementation by looking at the Debug variants, since then the optimization flags are consistent between Make and CMake, and indeed when the error is zero for Make then it is zero for CMake as well: see Linux/CMake/Debug/serial and Linux/CMake/Debug/dmpar. Does that make sense? Let us know if/how we should clarify this in the paper.

This is a very interesting and important result in itself, I think.  The implications of this comparison of different build pathways goes well beyond WRF-CMAKE, as your results show that even when sticking with upstream WRF, the choices I make when compiling do significantly affect my results.  Given the significance of this, I would strongly suggest the authors write a small *technical note* paper about this and publish it in modeling journals like GMD.

As for the present JOSS paper, in my eyes the paper would benefit if the authors would

1. better explain the issue with the NRSME differences (for example, using the explanation given here in the discussion thread)
2. explicitly mention how much WRF-CMAKE results differ from upstream WRF for each build configuration
