### 2019-08-17 c8fc15d
replace paper plots, wip [skip ci]
