+++ paper/paper.md
@@ -65,10 +65,75 @@ As noted by Hodyss and Majumdar [-@Hodyss2007], and Geer [-@Geer2016], the high 
 Table: WRF prognostic variables evaluated during integration tests.
 
 
+\begin{tiny}
+\begin{longtable}{llrrrrrr}
+\caption{Make vs CMake Domain 2, 60 min}\\
+\toprule
+                          &     &  $p$ in $\mathsf{Pa}$ &  $\phi$ in $\mathsf{m^2\ s^{-2}}$ &  $\theta$ in $\mathsf{K}$ &  $u$ in $\mathsf{m\ s^{-1}}$ &  $v$ in $\mathsf{m\ s^{-1}}$ &  $w$ in $\mathsf{m\ s^{-1}}$ \\
+Configuration & Statistic &                       &                                   &                           &                              &                              &                              \\
+\midrule
+\endhead
+\midrule
+\multicolumn{8}{r}{{Continued on next page}} \\
+\midrule
+\endfoot
+
+\bottomrule
+\endlastfoot
+Linux/Debug/dmpar & Mean &               0.0e+00 &                           0.0e+00 &                   0.0e+00 &                      0.0e+00 &                      0.0e+00 &                      0.0e+00 \\
+                    & SD &               0.0e+00 &                           0.0e+00 &                   0.0e+00 &                      0.0e+00 &                      0.0e+00 &                      0.0e+00 \\
+                    & Max &               0.0e+00 &                           0.0e+00 &                   0.0e+00 &                      0.0e+00 &                      0.0e+00 &                      0.0e+00 \\
+Linux/Debug/dm\_sm & Mean &               5.0e-04 &                           1.6e-02 &                   1.9e-04 &                      3.3e-04 &                      2.8e-04 &                      9.4e-05 \\
+                    & SD &               3.6e-03 &                           9.4e-02 &                   2.5e-03 &                      4.9e-03 &                      3.3e-03 &                      7.7e-04 \\
+                    & Max &               1.2e+00 &                           9.1e+00 &                   4.7e-01 &                      1.1e+00 &                      7.6e-01 &                      1.3e-01 \\
+Linux/Debug/serial & Mean &               0.0e+00 &                           0.0e+00 &                   0.0e+00 &                      0.0e+00 &                      0.0e+00 &                      0.0e+00 \\
+                    & SD &               0.0e+00 &                           0.0e+00 &                   0.0e+00 &                      0.0e+00 &                      0.0e+00 &                      0.0e+00 \\
+                    & Max &               0.0e+00 &                           0.0e+00 &                   0.0e+00 &                      0.0e+00 &                      0.0e+00 &                      0.0e+00 \\
+Linux/Debug/smpar & Mean &               2.9e-04 &                           1.3e-02 &                   1.4e-04 &                      1.8e-04 &                      1.7e-04 &                      7.6e-05 \\
+                    & SD &               2.8e-03 &                           1.0e-01 &                   1.7e-03 &                      1.6e-03 &                      1.5e-03 &                      8.1e-04 \\
+                    & Max &               1.2e+00 &                           1.2e+01 &                   4.7e-01 &                      2.6e-01 &                      2.5e-01 &                      1.7e-01 \\
+Linux/Release/dmpar & Mean &               3.0e-04 &                           1.3e-02 &                   1.3e-04 &                      1.6e-04 &                      1.6e-04 &                      8.0e-05 \\
+                    & SD &               3.0e-03 &                           1.1e-01 &                   1.8e-03 &                      1.4e-03 &                      1.4e-03 &                      8.3e-04 \\
+                    & Max &               1.2e+00 &                           1.1e+01 &                   4.6e-01 &                      2.1e-01 &                      2.6e-01 &                      1.3e-01 \\
[…]
