### 2019-09-01 b5a1390
update figures & adapt text [skip ci]
