### 2019-09-01 #1468 @letmaik
@andreas-h We took your feedback onboard and updated the figures and some text. This should hopefully provide a better overall picture, particularly the comparison of WRF upstream with WRF-CMake given the same build variant. Please have another look and let us know if this is all fine. Thanks!
