### 2019-09-02 #1468 @andreas-h
Thank's a lot for considering my concerns.

Once the TODO from Appendix B is sorted out, I'm happy to accept
the paper.
