### 2019-09-02 #1468 @letmaik
@andreas-h Thanks, that's great. The TODO from Appendix B is merely a typesetting issue. The figures are on the following page but should ideally be directly under the heading. @arfon Can you help with that? I think we're good otherwise.
