### 2019-09-02 #1468 @arfon
@letmaik - looks like the paper figure placement issue is resolved now.  Could you remove the TODO statement from the paper?
