### 2019-09-02 #1468 @arfon
@letmaik - could you also merge this PR: https://github.com/WRF-CMake/wrf/pull/39
