### 2019-09-02 fbabf64
minor paper improvements [skip ci]
