+++ paper/paper.md
@@ -49,11 +49,11 @@ Build tests are performed for all supported build variants (Table 1) using CI se
 
 Table: Build variants used in build and regression tests. `Make`: original WRF build system files, `CMake`: this paper; `Debug`: compiler optimizations disabled, `Release`: enabled; `serial`: single processor, `dmpar`: multiple with distributed memory (MPI), `smpar`: multiple with shared memory (OpenMP), `dm_sm`: multiple with MPI and OpenMP.
 
-As a result, the impact of our changes are evaluated using the range-normalized relative percentage error $\boldsymbol{\delta}$ and range-normalized root-mean-square percentage error (NRMSPE; Appendix A). These are computed per domain for all grid points, and for all vertical levels. The errors are assessed by (a) comparing the outputs of prognostic-variable outputs (Table 2) from WRF (Make) against those from WRF-CMake and (b) comparing the outputs for all build variants (for both Make and CMake) against a reference build variant defined as `Linux/Make/Debug/serial`.
+As a result, the impact of our changes are evaluated using the range-normalized relative percentage error ($\boldsymbol{\delta}_{x}$) and range-normalized root-mean-square percentage error (NRMSPE; Appendix A). These are computed per domain for all grid points, and for all vertical levels. The errors are assessed by (a) comparing the outputs of prognostic-variable outputs (Table 2) from WRF (Make) against those from WRF-CMake and (b) comparing the outputs for all build variants (for both Make and CMake) against a reference build variant defined as `Linux/Make/Debug/serial`.
 
-These tests are then run for all supported build variants (Table 1) using the [WRF-CMake Automated Testing Suite (WATS)](https://github.com/WRF-CMake/wats) and a subset of namelists[^2] from the official [WRF Testing Framework](https://github.com/wrf-model/WTF). These are run on CI services at major code changes (e.g. before merging pull requests), and for 1 hour of simulation time, to constrain computing resources.
+ These tests are then run for all supported build variants (Table 1) using the [WRF-CMake Automated Testing Suite (WATS)](https://github.com/WRF-CMake/wats), and a subset of namelists[^2] from the official [WRF Testing Framework](https://github.com/wrf-model/WTF), using CI services at major code changes (e.g. before merging pull requests), and for 1 hour of simulation time, to constrain computing resources.
 
-Here, we report the summary results for the domain showing the greatest error (i.e. innermost; domain 2) after simulating for 60 minutes. Values of $\boldsymbol{\delta}$ in the extended boxplots (Figure 1 and 3) are aggregated for all quantities reported in Table 2.
+Here, we report summary results for the domain showing the greatest error (i.e. innermost; domain 2) after simulating 60 minutes. Values of $\boldsymbol{\delta}_{x}$ are aggregated for all quantities reported in Table 2 and referred to as $\boldsymbol{\delta}$.
 
 | Symbol           | Name                                  | Unit                   |
 | ---------------- | ------------------------------------- | ---------------------- |
@@ -66,11 +66,11 @@ Here, we report the summary results for the domain showing the greatest error (i
 
 Table: WRF prognostic variables evaluated during regression tests.
 
[…]
