### 2019-09-02 34aa782
Update paper.bib (#39)
