+++ paper/paper.bib
@@ -50,7 +50,7 @@ year = {2017}
   publisher = {{IOP} Publishing},
   volume = {898},
   pages = {072010},
-  author = {J Elmsheuser and A Krasznahorkay and E Obreshkov and A Undrus and},
+  author = {J Elmsheuser and A Krasznahorkay and E Obreshkov and A Undrus},
   title = {Large Scale Software Building with {CMake} in {ATLAS}},
   journal = {Journal of Physics: Conference Series}
 }
