### 2019-09-04 f4b1a0a
switch to github download links for netcdf and jasper
