### 2019-09-04 ab9e62f
curl: follow redirects
