### 2019-09-07 #1468 @dmey
@arfon As Zenodo requires the software's list of authors to be included when depositing the software on their website, we were wondering what the best way is to clearly show that our work is merely an extension (build-system enhancement) of the great work carried out by Skamarock et al., (2018; and previous papers).

Have you had a similar case before or have any idea how to best show this on Zenodo? On GitHub this is perhaps easier to see given that the project is already marked as a fork of WRF and includes a statement under the [how to cite section in the README](https://github.com/WRF-CMake/wrf#how-to-cite) that links to the WRF model publication.

@zbeekman and @andreas-h if you also have any tips which you think could make this clearer, or think we should add anything else in the README etc, please let us know!
