### 2019-09-08 #1468 @arfon
> @arfon As Zenodo requires the software's list of authors to be included when depositing the software on their website, we were wondering what the best way is to clearly show that our work is merely an extension (build-system enhancement) of the great work carried out by Skamarock et al., (2018; and previous papers).

Hrm, good question. I this situation I think it's probably best to just include the authors that have extended the package in this fork.
