### 2019-09-09 #1468 @dmey
@arfon OK, thanks. I have added the following in the description on Zenodo to hopefully make it clear:

> As WRF-CMake is an extension of the original WRF model and software, we ask to please cite both model (i.e. Skamarock et al., (2018)), and software with version. For more information on how to do this correctly, please see the [How to cite section](https://github.com/WRF-CMake/wrf#how-to-cite) on the project's GitHub page.
